1218_ Text 개선 + 표현 가능한 문자 추가
Text 최적화
- 기존 Text의 문제점
  - 기존 Text는 각 Text 객체마다 모든 사용 가능한 문자에 대한 Frame 객체를 생성해서 소지
  - 거기에 더해 애니메이션이 없는데도 clip을 사용
- 개선 후
  - static unordered_map을 사용. 글자에 대한 Frame을 한 번만 생성하게 만듦.
  - 생성된 Text의 수를 저장하는 변수 추가. 해당 변수가 0이 되면 생성된 글자 Frame들을 제거
  - 초기화 여부를 관리하는 변수 추가. 해당 변수가 false일 경우에만 글자 frame 생성
  - clip 사용 없이 const Frame* 변수를 사용하도록 변경.

표현 가능한 문자 추가
- 기존 Text에서 - 와 '이 표현이 되지 않았던 것을 확인하여 수정을 진행

+++ DX_MyProject/Framework/Render/Clip.h
@@ -40,6 +40,6 @@ public:
 	void Stop();
 	
 	UINT GetFrameNum() { return cur_frame_num; }
-	UINT GetFrameCnt() { return frames.size(); }
+	size_t GetFrameCnt() { return frames.size(); }
 	void SetFrameNum(UINT num) { cur_frame_num = num; }
 };
